+++ internal/tlogclient/tlogclient.go
@@ -14,7 +14,7 @@ import (
 	"strings"
 	"time"
 
-	"filippo.io/litetlog/internal/tlogx"
+	"filippo.io/torchwood/internal/tlogx"
 	"golang.org/x/mod/sumdb/tlog"
 	"golang.org/x/sync/errgroup"
 )