+++ internal/tlogx/cosignature_test.go
@@ -5,7 +5,7 @@ import (
 	"crypto/rand"
 	"testing"
 
-	"filippo.io/litetlog/internal/tlogx"
+	"filippo.io/torchwood/internal/tlogx"
 	"golang.org/x/mod/sumdb/note"
 )